+++ Bot.GetByLink.Common.Infrastructure/MediaInfo.cs
@@ -27,7 +27,7 @@ public class MediaInfo : IMediaInfo
     /// <summary>
     ///     Gets or sets for size in mb on url media.
     /// </summary>
-    public long SizeMedia { get; set; }
+    public double SizeMedia { get; set; }
 
     /// <summary>
     ///     Function for set imfo media on url media.
